requirements.txt
openai
requests
beautifulsoup4
Pillow
streamlit>=1.30